Adding a link target from src attribute
This is a small hack. A didn't want to create a link (<a>) around all images, but just let the user click on an image to open lightcase.

To work, the lauch function has to be changed to :

jQuery(document).ready(function ($) {
    $('img[data-rel^=lightcase]').lightcase();
    $('img[data-rel^=lightcase]').css('cursor', 'pointer');
});

It also changes the cursor to pointer for all images with data-rel="lightcase".